feat: separate kftray and header component
